+++ test/directory.test.js
@@ -89,9 +89,9 @@ describe('directory', function() {
               id: '703887',
               username: 'mhashimoto',
               emails: [{
-                address: 'mhashimoto-04@plaxo.com'
+                value: 'mhashimoto-04@plaxo.com'
               }, {
-                address: 'mhashimoto@plaxo.com'
+                value: 'mhashimoto@plaxo.com'
               }]
             });
             done();
@@ -99,6 +99,51 @@ describe('directory', function() {
         })
         .catch(done);
     }); // should map emails with address
+    
+    it('should map phone numbers with number', function(done) {
+      var client = new Object();
+      client.query = sinon.stub();
+      client.query.onFirstCall().resolves(null);
+      client.query.onSecondCall().yieldsAsync(null, {
+        rows: [
+          {
+            user_id: '703887',
+            username: 'mhashimoto',
+            phone_numbers: '{"(KLONDIKE5,,,)","(650-123-4567,,,)"}'
+          }
+        ]
+      });
+      
+      var postgres = new Object();
+      postgres.createConnectionPool = sinon.stub().returns(client);
+    
+      var directory = factory('postgresql://www.example.com/exampledb', postgres)
+        .then(function(directory) {
+          expect(postgres.createConnectionPool).to.have.been.calledOnceWith('postgresql://www.example.com/exampledb');
+          
+          directory.read('703887', function(err, user) {
+            if (err) { return done(err); }
+        
+            expect(client.query).to.have.been.calledTwice;
+            var sql = client.query.getCall(1).args[0];
+            var values = client.query.getCall(1).args[1];
+            expect(sql).to.equal('SELECT *    FROM users   WHERE user_id = $1');
+            expect(values).to.deep.equal([ '703887' ]);
+        
+            expect(user).to.deep.equal({
+              id: '703887',
+              username: 'mhashimoto',
+              phoneNumbers: [{
+                value: 'KLONDIKE5'
+              }, {
+                value: '650-123-4567'
+              }]
+            });
+            done();
+          });
+        })
+        .catch(done);
+    }); // should map phone numbers with number
     
   }); // #read