+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+### Added
+- Parsing phone numbers for user object yielded from `PostgreSQLUserDirectory#read`.
+
 ## [0.0.5] - 2024-05-10
 ### Fixed
 - Fixed `ENOENT` error that occurs when attempting to create `clients` table.